Changed AI.minMax return value and disabled A-B pruning
AI.minMax now returns a TreeMap with a column as its key and the generated score as its value. This should make the function AI.bestColumn more accurate in terms of associating the generated scores with a column.

Alpha-Beta pruning is efficient but results in the AI making bad moves. This function would probably be disabled in the final release. An adapted version of A-B pruning is used if a player wins.

Signed-off-by: Vincent Guo <vincentguo04@gmail.com>